added drone data example and fixed issue for drone landing in start position.

+++ include/core/graphprocessor/graphprocessor.h
@@ -27,7 +27,7 @@ public:
     void startExploration();
     void startAttack();
     void setImgSize(size_t imgW, size_t imgH);
-    void sendFileToDrone(QString serverIP, quint16 port, QFile file);
+    void sendFileToDrone(QString serverIP, int port, QFile file);
 
 private:
